@@ -2,89 +2,96 @@
 namespace General;
 
 use Zend\Stdlib\ArrayUtils;
-
+use General\Controller\ControllerInitializer;
+use General\Service\ServiceInitializer;
 /**
  * ITEA Office copyright message placeholder
  *
  * @category    General
  * @package     Config
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright   Copyright (c] 2004-2014 ITEA Office (http://itea3.org]
  */
-$config = array(
-    'controllers'     => array(
-        'invokables' => array(
+$config = [
+    'controllers'     => [
+        'initializers' => [
+            ControllerInitializer::class
+        ],
+        'invokables' => [
             'general-index' => 'General\Controller\IndexController',
             'general-style' => 'General\Controller\StyleController',
-        ),
-    ),
-    'view_manager'    => array(
+        ],
+    ],
+    'view_manager'    => [
         'template_map' => include __DIR__ . '/../template_map.php',
-    ),
-    'service_manager' => array(
-        'factories'  => array(
+    ],
+    'service_manager' => [
+        'initializers' => [
+            ServiceInitializer::class
+        ],
+        'factories'    => [
             'general-assertion' => 'General\Acl\Assertion\General',
-        ),
-        'invokables' => array(
+        ],
+        'invokables'   => [
             'general_general_service' => 'General\Service\GeneralService',
             'general_form_service'    => 'General\Service\FormService',
 
-        )
-    ),
-    'asset_manager'   => array(
-        'resolver_configs' => array(
-            'collections' => array(
-                'assets/' . (defined("DEBRANOVA_HOST") ? DEBRANOVA_HOST : 'test') . '/js/jvectormap.js' => array(
+        ]
+    ],
+    'asset_manager'   => [
+        'resolver_configs' => [
+            'collections' => [
+                'assets/' . (defined("DEBRANOVA_HOST") ? DEBRANOVA_HOST : 'test') . '/js/jvectormap.js' => [
                     'js/jquery/jquery.mousewheel.min.js',
                     'js/jquery/jquery-jvectormap-1.1.1.min.js',
                     'js/jquery/jquery-jvectormap-europe-mill-en.js',
-                ),
-            ),
-            'paths'       => array(
+                ],
+            ],
+            'paths'       => [
                 __DIR__ . '/../public',
-            ),
-            'caching'     => array(
-                'assets/' . (defined("DEBRANOVA_HOST") ? DEBRANOVA_HOST : 'test') . '/js/jvectormap.js' => array(
+            ],
+            'caching'     => [
+                'assets/' . (defined("DEBRANOVA_HOST") ? DEBRANOVA_HOST : 'test') . '/js/jvectormap.js' => [
                     'cache'   => 'FilePath', //Filesystem for development
-                    'options' => array(
+                    'options' => [
                         'dir' => __DIR__ . '/../../../public',
-                    ),
-                ),
-            ),
-        ),
-    ),
-    'doctrine'        => array(
-        'driver'       => array(
-            'general_annotation_driver' => array(
+                    ],
+                ],
+            ],
+        ],
+    ],
+    'doctrine'        => [
+        'driver'       => [
+            'general_annotation_driver' => [
                 'class' => 'Doctrine\ORM\Mapping\Driver\AnnotationDriver',
-                'paths' => array(
+                'paths' => [
                     __DIR__ . '/../src/General/Entity/'
-                )
-            ),
-            'orm_default'               => array(
+                ]
+            ],
+            'orm_default'               => [
                 'class'   => 'Doctrine\ORM\Mapping\Driver\DriverChain',
-                'drivers' => array(
+                'drivers' => [
                     __NAMESPACE__ . '\Entity' => 'general_annotation_driver',
-                )
-            )
-        ),
-        'eventmanager' => array(
-            'orm_general' => array(
-                'subscribers' => array(
+                ]
+            ]
+        ],
+        'eventmanager' => [
+            'orm_general' => [
+                'subscribers' => [
                     'Gedmo\Timestampable\TimestampableListener',
                     'Gedmo\Sluggable\SluggableListener',
-                )
-            ),
-        ),
-    )
-);
+                ]
+            ],
+        ],
+    ]
+];
 
-$configFiles = array(
+$configFiles = [
     __DIR__ . '/module.config.routes.php',
     __DIR__ . '/module.config.general.php',
     __DIR__ . '/module.config.navigation.php',
     __DIR__ . '/module.config.authorize.php',
-);
+];
 
 foreach ($configFiles as $configFile) {
     $config = ArrayUtils::merge($config, include $configFile);
