@@ -1,46 +1,46 @@
 <?php
 // Generated by ZF2's ./bin/classmap_generator.php
 return array(
-  'General\Options\ModuleOptions'                => __DIR__ . '/src/General/Options/ModuleOptions.php',
-  'General\Repository\Country'                   => __DIR__ . '/src/General/Repository/Country.php',
-  'General\Entity\EntityAbstract'                => __DIR__ . '/src/General/Entity/EntityAbstract.php',
-  'General\Entity\VatType'                       => __DIR__ . '/src/General/Entity/VatType.php',
-  'General\Entity\Web'                           => __DIR__ . '/src/General/Entity/Web.php',
-  'General\Entity\Eureka'                        => __DIR__ . '/src/General/Entity/Eureka.php',
-  'General\Entity\Country'                       => __DIR__ . '/src/General/Entity/Country.php',
-  'General\Entity\Flag'                          => __DIR__ . '/src/General/Entity/Flag.php',
-  'General\Entity\HydrateInterface'              => __DIR__ . '/src/General/Entity/HydrateInterface.php',
-  'General\Entity\WebInfo'                       => __DIR__ . '/src/General/Entity/WebInfo.php',
-  'General\Entity\Title'                         => __DIR__ . '/src/General/Entity/Title.php',
-  'General\Entity\ContentType'                   => __DIR__ . '/src/General/Entity/ContentType.php',
-  'General\Entity\Challenge'                     => __DIR__ . '/src/General/Entity/Challenge.php',
-  'General\Entity\Itac'                          => __DIR__ . '/src/General/Entity/Itac.php',
-  'General\Entity\CommunityType'                 => __DIR__ . '/src/General/Entity/CommunityType.php',
-  'General\Entity\EntityInterface'               => __DIR__ . '/src/General/Entity/EntityInterface.php',
-  'General\Entity\Vat'                           => __DIR__ . '/src/General/Entity/Vat.php',
-  'General\Entity\Gender'                        => __DIR__ . '/src/General/Entity/Gender.php',
-  'General\Entity\Eu'                            => __DIR__ . '/src/General/Entity/Eu.php',
+  'General\View\Helper\ChallengeLink'            => __DIR__ . '/src/General/View/Helper/ChallengeLink.php',
+  'General\View\Helper\GeneralServiceProxy'      => __DIR__ . '/src/General/View/Helper/GeneralServiceProxy.php',
+  'General\View\Helper\CountryFlag'              => __DIR__ . '/src/General/View/Helper/CountryFlag.php',
+  'General\View\Helper\CountryMap'               => __DIR__ . '/src/General/View/Helper/CountryMap.php',
+  'General\View\Helper\ContentTypeIcon'          => __DIR__ . '/src/General/View/Helper/ContentTypeIcon.php',
+  'General\View\Helper\CountryHandler'           => __DIR__ . '/src/General/View/Helper/CountryHandler.php',
+  'General\View\Helper\CountryLink'              => __DIR__ . '/src/General/View/Helper/CountryLink.php',
+  'General\View\Helper\ChallengeHandler'         => __DIR__ . '/src/General/View/Helper/ChallengeHandler.php',
+  'General\Version\Version'                      => __DIR__ . '/src/General/Version/Version.php',
+  'General\Service\EmailService'                 => __DIR__ . '/src/General/Service/EmailService.php',
+  'General\Service\GeneralService'               => __DIR__ . '/src/General/Service/GeneralService.php',
+  'General\Service\ServiceAbstract'              => __DIR__ . '/src/General/Service/ServiceAbstract.php',
+  'General\Service\GeneralServiceAwareInterface' => __DIR__ . '/src/General/Service/GeneralServiceAwareInterface.php',
+  'General\Service\ServiceInitializer'           => __DIR__ . '/src/General/Service/ServiceInitializer.php',
+  'General\Service\ServiceInterface'             => __DIR__ . '/src/General/Service/ServiceInterface.php',
+  'General\Service\EmailServiceAwareInterface'   => __DIR__ . '/src/General/Service/EmailServiceAwareInterface.php',
+  'General\Service\FormService'                  => __DIR__ . '/src/General/Service/FormService.php',
+  'General\Service\FormServiceAwareInterface'    => __DIR__ . '/src/General/Service/FormServiceAwareInterface.php',
   'General\Controller\StyleController'           => __DIR__ . '/src/General/Controller/StyleController.php',
   'General\Controller\ControllerInitializer'     => __DIR__ . '/src/General/Controller/ControllerInitializer.php',
   'General\Controller\IndexController'           => __DIR__ . '/src/General/Controller/IndexController.php',
-  'General\View\Helper\GeneralServiceProxy'      => __DIR__ . '/src/General/View/Helper/GeneralServiceProxy.php',
-  'General\View\Helper\CountryMap'               => __DIR__ . '/src/General/View/Helper/CountryMap.php',
-  'General\View\Helper\CountryHandler'           => __DIR__ . '/src/General/View/Helper/CountryHandler.php',
-  'General\View\Helper\ChallengeHandler'         => __DIR__ . '/src/General/View/Helper/ChallengeHandler.php',
-  'General\View\Helper\ChallengeLink'            => __DIR__ . '/src/General/View/Helper/ChallengeLink.php',
-  'General\View\Helper\CountryLink'              => __DIR__ . '/src/General/View/Helper/CountryLink.php',
-  'General\View\Helper\ContentTypeIcon'          => __DIR__ . '/src/General/View/Helper/ContentTypeIcon.php',
-  'General\View\Helper\CountryFlag'              => __DIR__ . '/src/General/View/Helper/CountryFlag.php',
-  'General\Version\Version'                      => __DIR__ . '/src/General/Version/Version.php',
   'General\Module'                               => __DIR__ . '/src/General/Module.php',
-  'General\Service\EmailService'                 => __DIR__ . '/src/General/Service/EmailService.php',
-  'General\Service\GeneralService'               => __DIR__ . '/src/General/Service/GeneralService.php',
-  'General\Service\EmailServiceAwareInterface'   => __DIR__ . '/src/General/Service/EmailServiceAwareInterface.php',
-  'General\Service\ServiceInitializer'           => __DIR__ . '/src/General/Service/ServiceInitializer.php',
-  'General\Service\ServiceAbstract'              => __DIR__ . '/src/General/Service/ServiceAbstract.php',
-  'General\Service\ServiceInterface'             => __DIR__ . '/src/General/Service/ServiceInterface.php',
-  'General\Service\FormService'                  => __DIR__ . '/src/General/Service/FormService.php',
-  'General\Service\FormServiceAwareInterface'    => __DIR__ . '/src/General/Service/FormServiceAwareInterface.php',
-  'General\Service\GeneralServiceAwareInterface' => __DIR__ . '/src/General/Service/GeneralServiceAwareInterface.php',
+  'General\Entity\HydrateInterface'              => __DIR__ . '/src/General/Entity/HydrateInterface.php',
+  'General\Entity\WebInfo'                       => __DIR__ . '/src/General/Entity/WebInfo.php',
+  'General\Entity\Eu'                            => __DIR__ . '/src/General/Entity/Eu.php',
+  'General\Entity\ContentType'                   => __DIR__ . '/src/General/Entity/ContentType.php',
+  'General\Entity\EntityAbstract'                => __DIR__ . '/src/General/Entity/EntityAbstract.php',
+  'General\Entity\Itac'                          => __DIR__ . '/src/General/Entity/Itac.php',
+  'General\Entity\Web'                           => __DIR__ . '/src/General/Entity/Web.php',
+  'General\Entity\Country'                       => __DIR__ . '/src/General/Entity/Country.php',
+  'General\Entity\Gender'                        => __DIR__ . '/src/General/Entity/Gender.php',
+  'General\Entity\CommunityType'                 => __DIR__ . '/src/General/Entity/CommunityType.php',
+  'General\Entity\Eureka'                        => __DIR__ . '/src/General/Entity/Eureka.php',
+  'General\Entity\Flag'                          => __DIR__ . '/src/General/Entity/Flag.php',
+  'General\Entity\VatType'                       => __DIR__ . '/src/General/Entity/VatType.php',
+  'General\Entity\Vat'                           => __DIR__ . '/src/General/Entity/Vat.php',
+  'General\Entity\Title'                         => __DIR__ . '/src/General/Entity/Title.php',
+  'General\Entity\EntityInterface'               => __DIR__ . '/src/General/Entity/EntityInterface.php',
+  'General\Entity\Challenge'                     => __DIR__ . '/src/General/Entity/Challenge.php',
   'General\Email'                                => __DIR__ . '/src/General/Email.php',
+  'General\Repository\Country'                   => __DIR__ . '/src/General/Repository/Country.php',
+  'General\Options\ModuleOptions'                => __DIR__ . '/src/General/Options/ModuleOptions.php',
 );
