@@ -151,7 +151,7 @@ class Eureka extends EntityAbstract
      */
     public function getArrayCopy()
     {
-        return array();
+        return [];
     }
 
     public function populate()
