@@ -141,7 +141,7 @@ class Itac extends EntityAbstract
      */
     public function getArrayCopy()
     {
-        return array();
+        return [];
     }
 
     public function populate()
