@@ -1,15 +1,15 @@
 <?php
 // Generated by ZF2's ./bin/templatemap_generator.php
 return array(
+    'email/newsletter'                        => __DIR__ . '//view/email/newsletter.twig',
     'general/partial/list/country'            => __DIR__ . '//view/general/partial/list/country.twig',
-    'general/partial/list/challenge'          => __DIR__ . '//view/general/partial/list/challenge.twig',
-    'general/partial/list/article'            => __DIR__ . '//view/general/partial/list/article.twig',
-    'general/partial/list/organisation'       => __DIR__ . '//view/general/partial/list/organisation.twig',
+    'general/partial/list/project-challenge'  => __DIR__ . '//view/general/partial/list/project-challenge.twig',
     'general/partial/list/country-itac'       => __DIR__ . '//view/general/partial/list/country-itac.twig',
     'general/partial/list/project'            => __DIR__ . '//view/general/partial/list/project.twig',
-    'general/partial/list/project-challenge'  => __DIR__ . '//view/general/partial/list/project-challenge.twig',
+    'general/partial/list/organisation'       => __DIR__ . '//view/general/partial/list/organisation.twig',
+    'general/partial/list/article'            => __DIR__ . '//view/general/partial/list/article.twig',
+    'general/partial/list/challenge'          => __DIR__ . '//view/general/partial/list/challenge.twig',
     'general/partial/entity/country'          => __DIR__ . '//view/general/partial/entity/country.twig',
     'general/partial/entity/country-metadata' => __DIR__ . '//view/general/partial/entity/country-metadata.twig',
     'general/partial/entity/challenge'        => __DIR__ . '//view/general/partial/entity/challenge.twig',
-    'email/newsletter'                        => __DIR__ . '//view/email/newsletter.twig',
 );
