@@ -31,6 +31,6 @@ return array(
     ),
     'service_manager'         => array(
         'use_defaults' => true,
-        'factories'    => array(),
+        'factories'    => [],
     ),
 );
